Safari-611.1.10/LayoutTests/fast/canvas/webgl/webgl2-texStorage-expected.txt

Test that texStorage2D() works.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
asdf
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.checkFramebufferStatus(gl.FRAMEBUFFER) is gl.FRAMEBUFFER_COMPLETE
PASS gl.getError() is gl.NO_ERROR
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS receiver[i] is 0
PASS gl.getError() is not gl.NO_ERROR
PASS gl.getError() is not gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS receiver[i] is 1
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS successfullyParsed is true

TEST COMPLETE
